Design: 로그인 UI 개선
