+++ src/pages/Home.js
@@ -1,4 +1,5 @@
-import React from "react";
+import React, { useEffect } from "react";
+import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import { useQuery, useMutation, useQueryClient } from "react-query";
 import {
@@ -8,9 +9,16 @@ import {
   Item,
   InfiniteLanding,
 } from "../components";
-import { tweetAPI } from "../shared/api";
 
 const Mainpage = () => {
+  const navigate = useNavigate();
+  
+  useEffect(() => {
+    if (!localStorage.getItem("access_token")) {
+      navigate("/first");
+    }
+  }, [navigate]);
+
   return (
     <>
       <Header />
